Auth checck for edit and delete functions in posts show view. Will need to use more Materialize framework for styling. Good update point.

+++ app/views/posts/show.blade.php
@@ -1,9 +1,10 @@
 @extends('layouts.master')
 
 @section('content')
-    <div class="panel">
+<main>
+    <div class="container">
         <div class="row">
-            <div class="col-sm-8 blog-class">
+            <div class="col s12">
                 <div class="blog-post">
                     <h2 class="blog-title">{{{ $post->title }}}</h2>
                     <p class="blog-post-meta">{{{ $post->created_at->format('F-d-Y') }}}</p>
@@ -12,12 +13,25 @@
                 </div>
             </div>
         </div>
-        <ol class="breadcrumb">
-            <li><a href="{{{ action('PostsController@index') }}}">All Posts</a></li>
-            <li><a href="{{{ action('PostsController@edit', $post->id) }}}">Edit Post</a></li>
-        </ol>
-        {{ Form::open(['action' => ['PostsController@destroy', $post->id], 'method' => 'DELETE'])}}
-            <button class='btn btn-danger'>DELETE</button>
-        {{ FORM::close() }}
     </div>
+    <div class="container">
+        <ul class="breadcrumb">
+            <li>
+                <a href="{{{ action('PostsController@index') }}}">All Posts</a>
+            </li>
+        @if(Auth::check())
+            <li>
+                <a href="{{{ action('PostsController@edit', $post->id) }}}">Edit Post</a>
+            </li>
+        @endif
+        </ul>
+
+        @if(Auth::check())
+            {{ Form::open(['action' => ['PostsController@destroy', $post->id], 'method' => 'DELETE'])}}
+                <button class='btn btn-danger'>DELETE</button>
+            {{ FORM::close() }}
+        @endif
+    </div>
+    <br>
+</main>
 @stop
